refactor: rename types.ts to client.ts for better semantic clarity
