--- packages/tools/src/__tests__/index.test.ts
@@ -1,9 +1,0 @@
-describe('Main Index Exports', () => {
-  describe('Type Exports', () => {
-    it('should export Tool type', () => {
-      // The types are re-exported, so we can't test them at runtime
-      // The TypeScript compiler ensures they're properly exported
-      expect(true).toBe(true);
-    });
-  });
-});
