+++ packages/tools/src/__tests__/providers/openai/etf.test.ts
@@ -5,7 +5,7 @@ const mockETF = {
   getProfile: jest.fn(),
 };
 
-jest.mock('@/types', () => ({
+jest.mock('@/client', () => ({
   getFMPClient: jest.fn(() => ({
     etf: mockETF,
   })),
