+++ packages/tools/src/__tests__/providers/openai/insider.test.ts
@@ -4,7 +4,7 @@ const mockInsider = {
   getInsiderTradesBySymbol: jest.fn(),
 };
 
-jest.mock('@/types', () => ({
+jest.mock('@/client', () => ({
   getFMPClient: jest.fn(() => ({
     insider: mockInsider,
   })),
