+++ packages/tools/src/__tests__/providers/openai/market.test.ts
@@ -14,7 +14,7 @@ const mockMarket = {
   getMostActive: jest.fn(),
 };
 
-jest.mock('@/types', () => ({
+jest.mock('@/client', () => ({
   getFMPClient: jest.fn(() => ({
     market: mockMarket,
   })),
