+++ packages/tools/src/__tests__/providers/openai/senate-house.test.ts
@@ -16,7 +16,7 @@ const mockSenateHouse = {
   getHouseTradingRSSFeed: jest.fn(),
 };
 
-jest.mock('@/types', () => ({
+jest.mock('@/client', () => ({
   getFMPClient: jest.fn(() => ({
     senateHouse: mockSenateHouse,
   })),
