+++ packages/tools/src/__tests__/providers/openai/stock.test.ts
@@ -6,7 +6,7 @@ const mockStock = {
   getDividendHistory: jest.fn(),
 };
 
-jest.mock('@/types', () => ({
+jest.mock('@/client', () => ({
   getFMPClient: jest.fn(() => ({
     stock: mockStock,
   })),
