+++ packages/tools/src/__tests__/providers/vercel-ai/economic.test.ts
@@ -5,7 +5,7 @@ const mockEconomic = {
   getEconomicIndicators: jest.fn(),
 };
 
-jest.mock('@/types', () => ({
+jest.mock('@/client', () => ({
   getFMPClient: jest.fn(() => ({
     economic: mockEconomic,
   })),
