+++ packages/tools/src/__tests__/providers/vercel-ai/financial.test.ts
@@ -7,7 +7,7 @@ const mockFinancial = {
   getFinancialRatios: jest.fn(),
 };
 
-jest.mock('@/types', () => ({
+jest.mock('@/client', () => ({
   getFMPClient: jest.fn(() => ({
     financial: mockFinancial,
   })),
