+++ packages/tools/src/__tests__/providers/vercel-ai/institutional.test.ts
@@ -4,7 +4,7 @@ const mockInstitutional = {
   getInstitutionalHolders: jest.fn(),
 };
 
-jest.mock('@/types', () => ({
+jest.mock('@/client', () => ({
   getFMPClient: jest.fn(() => ({
     institutional: mockInstitutional,
   })),
