+++ packages/tools/src/__tests__/providers/vercel-ai/market.test.ts
@@ -4,7 +4,7 @@ const mockMarket = {
   getMarketPerformance: jest.fn(),
 };
 
-jest.mock('@/types', () => ({
+jest.mock('@/client', () => ({
   getFMPClient: jest.fn(() => ({
     market: mockMarket,
   })),
