+++ packages/tools/src/__tests__/providers/vercel-ai/senate-house.test.ts
@@ -7,7 +7,7 @@ const mockSenateHouse = {
   getHouseTradingRSSFeed: jest.fn(),
 };
 
-jest.mock('@/types', () => ({
+jest.mock('@/client', () => ({
   getFMPClient: jest.fn(() => ({
     senateHouse: mockSenateHouse,
   })),
