+++ packages/tools/src/__tests__/providers/vercel-ai/stock.test.ts
@@ -6,7 +6,7 @@ const mockStock = {
   getDividendHistory: jest.fn(),
 };
 
-jest.mock('@/types', () => ({
+jest.mock('@/client', () => ({
   getFMPClient: jest.fn(() => ({
     stock: mockStock,
   })),
