--- packages/tools/src/index.ts
@@ -1,4 +1,0 @@
-// Main exports for fmp-ai-tools package
-
-// Re-export only the tools-specific types
-export type { Tool } from './types';
