+++ packages/tools/src/providers/openai/company.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod';
 import { createOpenAITool } from '@/utils/openai-tool-wrapper';
-import { getFMPClient } from '@/types';
+import { getFMPClient } from '@/client';
 
 export const getCompanyProfile = createOpenAITool({
   name: 'getCompanyProfile',
