+++ packages/tools/src/providers/openai/etf.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod';
 import { createOpenAITool } from '@/utils/openai-tool-wrapper';
-import { getFMPClient } from '@/types';
+import { getFMPClient } from '@/client';
 
 // Input schema for ETF holdings with optional date
 const etfHoldingsInputSchema = z.object({
