+++ packages/tools/src/providers/openai/insider.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod';
 import { createOpenAITool } from '@/utils/openai-tool-wrapper';
-import { getFMPClient } from '@/types';
+import { getFMPClient } from '@/client';
 
 // Input schema for insider trading with symbol and optional page
 const insiderTradingInputSchema = z.object({
