+++ packages/tools/src/providers/openai/quote.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod';
 import { createOpenAITool } from '@/utils/openai-tool-wrapper';
-import { getFMPClient } from '@/types';
+import { getFMPClient } from '@/client';
 
 // Input schema for stock quote with symbol
 const stockQuoteInputSchema = z.object({
