+++ packages/tools/src/providers/openai/senate-house.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod';
 import { createOpenAITool } from '@/utils/openai-tool-wrapper';
-import { getFMPClient } from '@/types';
+import { getFMPClient } from '@/client';
 
 // Input schema for symbol-based trading data
 const symbolInputSchema = z.object({
