+++ packages/tools/src/providers/vercel-ai/calendar.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod';
 import { createTool } from '@/utils/aisdk-tool-wrapper';
-import { getFMPClient } from '@/types';
+import { getFMPClient } from '@/client';
 
 export const calendarTools = {
   getEarningsCalendar: createTool({
